Add support for slow dispatch
