+++ tt_metal/distributed/distributed.cpp
@@ -13,9 +13,11 @@ void AddProgramToMeshWorkload(MeshWorkload& mesh_workload, Program&& program, co
 }
 
 void EnqueueMeshWorkload(MeshCommandQueue& mesh_cq, MeshWorkload& mesh_workload, bool blocking) {
-    mesh_workload.compile(mesh_cq.device());
-    mesh_workload.load_binaries(mesh_cq);
-    mesh_workload.generate_dispatch_commands(mesh_cq);
+    if (mesh_cq.device()->using_fast_dispatch()) {
+        mesh_workload.compile(mesh_cq.device());
+        mesh_workload.load_binaries(mesh_cq);
+        mesh_workload.generate_dispatch_commands(mesh_cq);
+    }
     mesh_cq.enqueue_mesh_workload(mesh_workload, blocking);
 }
 
@@ -36,6 +38,9 @@ MeshEvent EnqueueRecordEventToHost(
 void EnqueueWaitForEvent(MeshCommandQueue& mesh_cq, const MeshEvent& event) { mesh_cq.enqueue_wait_for_event(event); }
 
 void EventSynchronize(const MeshEvent& event) {
+    if (event.device()->using_slow_dispatch()) {
+        return;
+    }
     for (const auto& coord : event.device_range()) {
         auto physical_device = event.device()->get_device(coord);
         while (physical_device->sysmem_manager().get_last_completed_event(event.mesh_cq_id()) < event.id());
